Refactoring Neuron Net.
