+++ src/digrec/stages/stage5/Recognation.java
@@ -1,0 +1,5 @@
+package digrec.stages.stage5;
+
+public class Recognation {
+
+}
